HEL-15 | Add the initial Docker configuration
Added an initial Docker configuration so that the application can be
run using docker-compose up. The setup will start a PostgreSQL database
and a Django backend.

I changed the settings.py so that most of the relevant configurations
are now read from the environment variables which can be entered using
docker-compose.env.yaml. I added an example file for reference.

I updated the requirements since I had to install django-environ for
handling the environment variables.

I added a custom command for creating an admin user and added
functionality to docker-entrypoint.sh so that the admin password is
either read from environment or generated randomly.

I modified the existing command for creating initial page contents so
that it can be run multiple times without error. I also added some
new page content initialization so that the front page will display
at least something instead of an error. I also added initial contents
for the welcome popup.

I had to modify the UserProfile model so that printer_ip is now a
GenericIPAddress field instead of IPAddress. The page content creation
would not start unless I made this change. GenericIPAddress should be
compatible with IPAddress using PostgreSQL, so this shouldn't be a
problem.

I added Docker description to README.md + added CONTRIBUTING.md so
that the documentation is on par with our practices.