HEL-16 | Add Gitlab configuration
Added the gitlab config file which points to the on_deploy.sh script
which will be run after deployment. I configured the script to perform
the migration, create an admin user and add the initial page contents.

I refactored the original admin user creation which was in the Docker
entrypoint script so that the add_admin_user command does the actual
password generation instead of the script. This way there's slightly
less code duplication.