HEL-127 | Change showcase owner display logic
Change the way the front page determines what is displayed as the owner of
a showcased collection. Earlier this used the is_featured flag of a Collection,
but the more logical solution is to do this automatically by checking if the
owner of the displayed Collection is a staff member or not. If they are staff,
the Helsinki logo is displayed. Otherwise the user icon + username.

I also added some prefetching in the queries to lessen the number of db
queries on the frontpage.