HEL-174 | Fix the collection view's slowness
When a Collection view was opened, the app would just get the Collection's
Records from the db and display the view. The view called Record.get_details()
for each Record which resulted in as many Finna queries which caused the view
to be slow. The results of each query were stored in the cache, so after the
initial hit the view was actually quite fast.

I refactored this so that the app checks which of the Collection's Records are
not in the cache and queries the missing Records from Finna using a single
query. The Finna results are then inserted into the cache so that the rest of
the view logic functions as it used to.

I also noticed that the app would still query Finna again when a single Record
was opened. I refactored this so that the Finna result will be picked up using
the same Record.get_details() so that the result should come from the cache or
at least fetched and put in the cache.

I also found out that the popover-info view was not used anymore even though it
was included as a hidden element which was displaying Finna details. It's been
replaced by the "scroll down" feature of the page, so I just removed it from the
app.