HEL-174 | Stop using the cache in the collections view
After some testing I came to the conclusion that using the cache is not
necessary in the collections view. Querying all of the collection's records
from Finna using a single query is fast enough, so I refactored the solution
so that Finna is queried every time a collection's details are opened.

This gives us a simpler logic for the view and we can stop using the cache
which is anyway a bit unreliable in its current configuration.