HEL-188 | Add GCS configuration
I added the necessary configuration for getting the staging environment to
work with GCS. I tested this locally and was able to fix the previous issues
when saving images using InMemoryUploadedFile. Turned out that when the image
was written into memory, a StringIO buffer was used so that the buffer was
pointing at the end of the image which in turn was interpreted by the storage
library so that there was no image to read. I fixed this by including seek(0)
statements before the InMemoryUploadedFile was created.

I also added a quick fix for the "crop and download" UI so that the download
is at least started. The browser is now handling it as a popup, since I used
window.open() to get it going. This should probably be fixed at some point.

I also added the environment variable configuration for Azure storage, but
it's currently missing a working storage library. At least locally it wasn't
working yet.