HEL-194 | Remove obsolete image fields from Record
I removed the two image fields from Record. The fields were only used in the
"Add image to collection" functionality which fetched two separate copies of
the selected image and stored them locally. In our recent changes we've
removed all features which did something with the local copies, so it was
unnecessary to keep the functionality in our codebase, let alone our file
system.

Since the "Add to collection" doesn't really do anything with cropped images,
I changed the title of the modal view so that it doesn't mention cropping.

I however left the old cropping functionality on the modal window, since I couldn't figure it out quickly. I created HEL-206 for finishing that later.