HEL-54 | Stop repeating favorites queries in collection view
When the app is displaying a collections' records list, it needs to know
what records are marked as a favorite by the currently logged in user.
Earlier this was done in the template by calling `Record.is_favorite()`
which then made database queries. This resulted in 2*(num of records)
queries for every collection details view. The queries are quite fast
so the user would not experience a lot of delay because of them, but
they were still unnecessary.

I changed this now so that the favorites are fetched only once in the
view's setup phase. The view marks the listed records with `is_favorite`
which is then used by the template.

I also removed the now-obsolete `Record.is_favorite()` to prevent this from
happening again.