HEL-208 | Make analytics usage configurable with env variables
Earlier the app's Google Analytics usage was implemented in a way which did
not take the current environment into account, i.e. every development, test,
review etc environment was sending analytics data into Google, therefore
making a mess of the actual production data.

I added a new environment variable `ENABLE_ANALYTICS` which can now be used
to control whether to include the GA scripts or not on the site. The default
value is False. The CI config is set to enable the analytics in production.