HEL-232 | Prevent invalid Finna queries
The previous production version of the app was referring to single images as
`(base url)/search/record/?search=some-search-term&page=123` which meant that
the app would do a Finna query with `some-search-term`, pick the 123rd result
and display it to the user.

Now that we've modified the search logic, these links are invalid. They are
however still used, probably in some Google caches or similar, so we've been
getting a lot of such requests in production since launching the service in
Kuva infra.

Our modified search was still using the `/search/record` URL, but it was
expecting an `image_id` parameter which was not being given by the old-style
links. This caused the app to query Finna without the record id, resulting in
a 400 error from Finna.

I've now changed the app's logic so that the search results are displayed using
`/search/details`. This causes our app to respond with a 404 to the old-style
requests which should (hopefully) cause Google's caches and all the other
places to invalidate their links.

I've also added a check to the Finna logic so that if it somehow gets an empty
record id as parameter, it won't query Finna.