HEL-233 | Add support for legacy image urls
There are currently at least some users and/or services which use the older
versions of image urls to point to image details hosted by our service. We
want to support those urls so the users' links won't get broken.

This PR adds backwards support for the older urls so that the incoming
request redirects to the newer version of the url using a permanent redirect
(301). Basically `/record/<finna_id>/` gets translated to
`/search/details/?image_id=<finna_id>`.