HEL-240 | Configure extra data in log entries
Earlier our logging configuration only logged the basic message for each log
entry. Many log entries in the app however are adding extra data into each
entry using the `extra` dictionary parameter, but we couldn't add that data
into the log because those entries without the extra data would crash with a
`KeyError`.

I added a new log filter which ensures that the logged `LogRecord` will
always have the property `data` which is now used in the log formatter so
that we can see the extra data in the logs.