HEL-238 | Remove duplicate collections from "browse albums" view
"Browse albums" used to fetch all public collections which it would display
on the view as users' public albums. The featured albums were separated into
their own list by picking collections marked with `is_featured`. These were
then displayed separately on the view.

I changed this so that the users' public collections do not contain those
which are marked with `is_featured` so that the two lists do not contain
duplicate entries.