HEL-224 | Log outcomes of login attempts
The app was not creating any log for login attempts which might make it more
vulnerable to misuse.

I added some signal handlers for login events so that they are now logged to
console whose output is in turn sent to a centralized service where the log
entries may be viewed.

The following levels are used:
- failed login: warning
- succeeded login: info
- logout: debug

The configuration is set to log INFO and above.