HEL-42 | Show cropped images in shopping basket view
Earlier the shopping basket view was displaying the actual full resolution
images from Finna for each line in the order. This meant that the browser
fetched several megabytes from Finna only to display the images as small
thumbnails. In addition, the thumbnails contained the whole image and not
the actual cropped part which the user had selected, probably causing a lot
of confusion about what the user was ordering.

The cropped image's url was already made available when the basket was
displayed, so I just changed the view to use the cropped image's url
instead of the Finna url.