HEL-125 | Perform login after successful password reset
During the testing of this feature we noticed that even if the user reset
their password, they had no way of knowing what their original username was.
Initially we decided to fix this by allowing the users to log in using either
a username or an email address, but that led to new problems as the app
allowed non-unique emails + emails as usernames.

Instead of just refactoring the way we handle user registrations, we
decided to keep the old way of registering so that the user is logged in
automatically after they reset their password. This way the user gets access
to the site by knowing only their email address (which was the primary goal)
and they can see their username from the header (if they realize it's their
username).

So this change only adds the logging in after reset + some text changes to
reflect this.