HEL-263 | Clear removal notification timestamp after login
When we've sent a notification to a user telling that their account will be
deleted after one month, the user can "cancel" the scheduled deletion by
logging in to the service.

I added a signal handler which sets a user's `removal_notification_sent`
field to null after login. This will affect the scheduled removal so that the
user won't be considered for deletion until 11 months later, if they don't
login before that. At that point, they will get a new notification and their
`removal_notification_sent` is set again.