HEL-263 | Add new management command for sending removal notifications
Added the new management command `send_removal_notifications` which will
send an email notification to users who have not logged in in a while (=
335 days by default or a custom value if given as a parameter) and who
have not yet received a notification. Also added tests for the command.

I changed the staging environment's config so that the new command will run
before `clean_unused_data`.