HEL-263 | Add max duration to staging's cronjob
The initial amount of email notifications in production is estimated to be
~1500. Sending that many emails takes over 5 minutes at least in staging
which means that the default value for max duration (5 minutes) is too low.

We want to keep the job running to avoid errors and duplicate emails which
is what is happening in staging, so I bumped this up to 15 minutes.