HEL-288 | Skip invalid emails when sending removal notifications
We currently have some invalid email addresses in our user database and
we are not verifying the new users' emails, so we're bound to get more typos
in emails in the future. These invalid addresses caused the scheduled
notification run to crash when it encountered such an address, therefore
stopping the whole notification process.

I added a check which makes sure that when Mailgun responds with a 400 Bad
request, the run will just mark that user as having been sent the message
and continues with the run. If we have a wrong address for a user, then they
are anyway unable to request for a password reset and so their user account
can't be accessed.

I also added a test for this + changed the commands status printouts a bit.