HEL-289 | Rename Bambora to Visma Pay
Visma bought Bambora and they have now changed Bambora's name to Visma Pay.
This meant that we must change our app so that there are no references to
Bambora.

I changed the few localization texts which referred to Bambora, replaced the
banner in the checkout view with Visma Pay's own image and changed some of
the comments in the code so that Bambora is no longer mentioned.

I also got rid of a lot of the commented out localization lines which were
probably left there after implementing the PageContent things.