HEL-292 | Remove printer functionality and dependency to paramiko
I removed the whole class which was handling communication with the printer,
along with the bit in the checkout process which was calling the class.

The whole printer functionality has long since been discarded, as it was
considered to be troublesome and difficult to use. Originally it was meant
to give museum visitors the chance to use a dedicated computer in the museum's
premises to browse Helsinkikuvia.fi and order printouts which would be
printed immediately using a printer in the premises.

Since we have so many warnings etc regarding the various libs paramiko is
pulling in, I decided to remove printing completely.