HEL-298 | Fix typo in default front page image url
When there are no "show on front page" collections, as was the case in
production a while ago, the app shows a default image. For some reason
we had a typo in the image's url and we noticed it only in production
after the frontpage collections were removed via a scheduled.

I fixed the typo, so now the default image works in case there is a
similar issue in production.