HEL-299 | Add env flag for feedback congestion message
Originally the "feedback is congested" message was hardcoded into the app
as a quick way of putting the message on display. Now the feedback processing
speed is back to normal and the message is no longer needed, but we decided
to keep the message and control its visibility via an environment variable.

I've now added the `ENABLE_FEEDBACK_CONGESTION_MSG` variable which is a
boolean which controls whether the congestion message is displayed under the
feedback form or not. In the future, should a similar congestion arise, we
can just flip the env var and redeploy to get the message back out there.