Removed all the buying, ordering and checkout features
HEL-313. The pictures printing company Printmotor wants to end the contract. The possibility for customer to buy actual physical prints from the service needs to be disabled and it is said that it should never be reactivated, because people can still download the pictures and go to print them by themselves.