Use cookie for language selection
Using the session for language selection
is somewhat a deprecated feature in Django.

LanguageView inherits some logic from django.views.i18n.set_language
for setting the language cookie which will be used
by the LocaleMiddleware.

Language stored in the UserProfile wasn't really working properly
in LanguageMiddleware so it is now removed.

https://docs.djangoproject.com/en/3.0/releases/3.0/#miscellaneous

Refs HEL-311